Move all components into a panel
