--- src/App.css
@@ -1,22 +1,0 @@
-#app-content {
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  column-gap: 1rem;
-}
-
-#app-content-container {
-  width: 700px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-}
-
-#playback-buttons {
-  display: flex;
-  gap: 20px;
-}
-
