+++ src/App.module.css
@@ -1,0 +1,6 @@
+.content {
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
