+++ src/components/mainPanel/MainPanel.module.css
@@ -1,0 +1,23 @@
+.layout {
+    display: flex;
+    flex-direction: row;
+}
+
+.panel {
+    width: 1000px;
+    height: 700px;
+    background: var(--color-white);
+    border: 1px solid var(--color-light-grey);
+    border-radius: 10px;
+    box-shadow: var(--shadow);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 20px;
+}
+
+.buttons {
+    display: flex;
+    gap: 20px;
+}
